Change how levels work completely - introduce AttackWaves which contain a group of objects (enemy spawners, events, barriers...) to be spawned after some delay, don't move the player or the camera and instead stay near world origin

+++ AstroDroids/Coroutines/WaitForSeconds.cs
@@ -7,7 +7,7 @@ namespace AstroDroids.Coroutines
         DateTime startTime;
         TimeSpan waitTime;
 
-        public WaitForSeconds(float seconds)
+        public WaitForSeconds(double seconds)
         {
             waitTime = TimeSpan.FromSeconds(seconds);
             startTime = DateTime.Now;
